move pareto.py back to utils
